requirements.txt
# Core
anthropic>=0.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0

# Cache
redis>=5.0.0

# Scheduler
apscheduler>=3.10.0

# HTTP
httpx>=0.27.0

# Config
python-dotenv>=1.0.0

# Phase 1
pytrends>=4.9.2

# Auth
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
email-validator>=2.1.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
